requirements.txt
jupyterlab
openai-whisper
setuptools-rust
deepl
streamlit
watchdog
